backend/requirements.txt
anyio==4.9.0
astroid==2.15.8
autopep8==2.3.2
black==23.3.0
certifi==2025.1.31
click==8.1.8
dill==0.3.9
fastapi==0.95.2
flake8==6.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.1.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pydantic==1.10.21
pyflakes==3.0.1
pylint==2.17.4
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.27.0
tomlkit==0.13.2
typing_extensions==4.13.1
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
